+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 Please see [CONTRIBUTING.md](https://github.com/cucumber/cucumber/blob/master/CONTRIBUTING.md) on how to contribute to Cucumber.
 
 ## [Unreleased]
+
+## [8.10.0] - 2022-12-27
 ### Added
 - Add support for YAML as a configuration file format ([#2199](https://github.com/cucumber/cucumber-js/pull/2199))
 
@@ -1457,7 +1459,8 @@ this.Given(), this.When(), this.Then() and this.defineStep() ([#2](https://githu
 
 ## 0.0.1
 
-[Unreleased]: https://github.com/cucumber/cucumber-js/compare/v8.9.1...HEAD
+[Unreleased]: https://github.com/cucumber/cucumber-js/compare/v8.10.0...HEAD
+[8.10.0]: https://github.com/cucumber/cucumber-js/compare/v8.9.1...v8.10.0
 [8.9.1]: https://github.com/cucumber/cucumber-js/compare/v8.9.0...v8.9.1
 [8.9.0]: https://github.com/cucumber/cucumber-js/compare/v8.8.0...v8.9.0
 [8.8.0]: https://github.com/cucumber/cucumber-js/compare/v8.7.0...v8.8.0